+++ src/oop_ext/foundation/callback/__init__.py
@@ -1,11 +1,25 @@
 from ._callbacks import Callbacks
 from ._callback import Callback
+from ._typed_callback import (
+    Callback0,
+    Callback1,
+    Callback2,
+    Callback3,
+    Callback4,
+    Callback5,
+)
 from ._priority_callback import PriorityCallback
 from ._shortcuts import After, Before, Remove, WrapForCallback
 
 __all__ = [
     "Callbacks",
     "Callback",
+    "Callback0",
+    "Callback1",
+    "Callback2",
+    "Callback3",
+    "Callback4",
+    "Callback5",
     "PriorityCallback",
     "After",
     "Before",